리딤코드 상세 페이지 리팩토링 (#2617)
각 리딤코드의 copied 상태를 따로 관리하기 위해 컴포넌트로 분리함
